Organized scene_viewer into plugins for reuse and organization.
In attempt to make more complicated 3D examples with scene_viewer functionality more phesable without code duplication.
