+++ examples/tools/scene_viewer_utils/mod.rs
@@ -1,0 +1,5 @@
+mod camera_controller_plugin;
+mod scene_viewer_plugin;
+
+pub use camera_controller_plugin::*;
+pub use scene_viewer_plugin::*;
